requirements.txt
astroid==2.13.3
attrs==22.2.0
Automat==22.10.0
autopep8==2.0.1
behave==1.2.6
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
dill==0.3.6
dnspython==2.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
filelock==3.9.0
flake8==6.0.0
greenlet==2.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
isort==5.12.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
lazy-object-proxy==1.9.0
lxml==4.9.2
mccabe==0.7.0
numpy==1.24.1
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
parse==1.19.0
parse-type==0.6.0
parsel==1.7.0
platformdirs==2.6.2
playwright==1.29.0
pluggy==1.0.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
PyDispatcher==2.0.6
pyee==9.0.4
pyflakes==3.0.1
PyHamcrest==2.0.4
pylint==2.15.10
pymongo==4.3.3
pyOpenSSL==22.1.0
pytest==7.2.1
python-dateutil==2.8.2
python-scrapyd-api==2.1.2
pytz==2022.7
PyYAML==6.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.1
scrapy-playwright==0.0.24
scrapyd==1.3.0
scrapyd-client==1.2.2
service-identity==21.1.0
six==1.16.0
tldextract==3.4.0
tomli==2.0.1
tomlkit==0.11.6
Twisted==22.10.0
typing_extensions==4.4.0
uberegg==0.1.1
urllib3==1.26.13
w3lib==2.1.1
wrapt==1.14.1
zope.interface==5.5.2
